Fix all unit tests and linter errors
