+++ .github/workflows/lint-optional.yaml
@@ -1,0 +1,31 @@
+name: Lint (optional)
+
+env:
+  HOPR_GITHUB_REF: ${{ github.ref }}
+  RUST_BACKTRACE: "1"
+  SCCACHE_GHA_ENABLED: "true"
+  RUSTC_WRAPPER: "sccache"
+
+on:
+  pull_request:
+    types:
+      - synchronize
+      - ready_for_review
+    paths-ignore:
+      - 'docs/**/*'
+      - '**/*.md'
+concurrency:
+  group: ${{ github.head_ref }}-lint
+  cancel-in-progress: true
+jobs:
+  shellcheck:
+    name: Shellcheck
+    runs-on: ubuntu-2-core
+    if: github.event.pull_request.draft == false
+    steps:
+      - name: Checkout hoprnet repository
+        uses: actions/checkout@v3
+
+      - name: Check shell scripts (not enforced for now)
+        continue-on-error: true
+        run: shellcheck -a --norc -s bash -x scripts/*.sh
