+++ packages/core-ethereum/src/index.ts
@@ -1,6 +1,6 @@
 import { Multiaddr } from '@multiformats/multiaddr'
 import type { PeerId } from '@libp2p/interface-peer-id'
-import { ChainWrapper, createChainWrapper, Receipt } from './ethereum.js'
+import { ChainWrapper, createChainWrapper, Receipt, type DeploymentExtract } from './ethereum.js'
 import {
   AcknowledgedTicket,
   Balance,
@@ -36,7 +36,6 @@ import {
 import Indexer from './indexer/index.js'
 import { EventEmitter } from 'events'
 import type { IndexerEvents } from './indexer/types.js'
-import { DeploymentExtract } from './utils/utils.js'
 
 const log = debug('hopr-core-ethereum')
 
@@ -697,4 +696,4 @@ export default class HoprCoreEthereum extends EventEmitter {
 // export { useFixtures } from './indexer/index.mock.js'
 export { sampleChainOptions } from './ethereum.mock.js'
 
-export { ChannelEntry, Indexer, ChainWrapper, createChainWrapper }
+export { ChannelEntry, Indexer, ChainWrapper, createChainWrapper, DeploymentExtract, Ethereum_Hash }
