+++ packages/core/crates/core-network/src/network.rs
@@ -4,7 +4,7 @@ use std::time::Duration;
 
 use libp2p_identity::PeerId;
 
-use utils_log::{error, info, warn};
+use utils_log::{info, warn};
 use utils_metrics::metrics::{MultiGauge, SimpleGauge};
 
 #[cfg(feature = "wasm")]
